requirements.txt
# BRET Package Dependencies
# Core scientific computing
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning (for evaluation metrics)
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration
pyyaml>=5.4.0

# Optional: Development dependencies
# pytest>=6.2.0
# pytest-cov>=2.12.0
# black>=21.0
# flake8>=3.9.0
